+++ library/src/main/java/com/pengrad/telegrambot/request/GetUserChatBoosts.java
@@ -4,7 +4,7 @@ import com.pengrad.telegrambot.response.UserChatBoostsResponse;
 
 public class GetUserChatBoosts extends BaseRequest<GetUserChatBoosts, UserChatBoostsResponse> {
 
-    public GetUserChatBoosts(Object chatId, Integer userId) {
+    public GetUserChatBoosts(Object chatId, Long userId) {
         super(UserChatBoostsResponse.class);
         add("chat_id", chatId).add("user_id", userId);
     }
